pyproject.toml
[project]
name = "pageindex"
version = "0.1.0"
description = "Vectorless, reasoning-based RAG using hierarchical document indexing with Vertex AI"
license = "MIT"
keywords = [
    "pageindex",
    "rag",
    "document",
    "pdf",
    "markdown",
    "tree-index",
    "vertex-ai",
    "gemini",
    "reasoning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
readme = "README.md"
authors = [
    { name = "Soham Dutta", email = "sodutta@redhat.com" },
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "google-cloud-aiplatform>=1.38.0",
    "pymupdf>=1.24.0",
    "pypdf2>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
]

[project.urls]
homepage = "https://gitlab.corp.redhat.com/sodutta/pageindex"
repository = "https://gitlab.corp.redhat.com/sodutta/pageindex"

[project.scripts]
pageindex = "pageindex.cli:app"

[project.optional-dependencies]
docling = [
    "docling>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
all = [
    "pageindex[docling,dev]",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pageindex = ["*.yaml"]

[tool.ruff]
target-version = "py310"
line-length = 100
respect-gitignore = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
